refactor: split the publish module by stage of the publish path
The publish module carried the outgoing message, the layer pipeline, the
transform stacks, the typed publisher, the reply wirings and the transaction
scopes in one file. Each stage moves to its own module behind the same public
paths; the fields the neighbouring stages reach for become visible within the
module rather than to the whole crate.
